Move roundRect adj guide test into PptSlidesTest
Relocates testWriterEmitsAdjGuideForRoundRect from AutoShapeTest to
PptSlidesTest per maintainer feedback on PR #880, using the shared
zip/XML assertion helpers instead of manual ZipArchive/regex parsing.
