+++ tests/PhpPresentation/Tests/Shape/AutoShapeTest.php
@@ -20,15 +20,10 @@ declare(strict_types=1);
 
 namespace PhpOffice\PhpPresentation\Tests\Shape;
 
-use PhpOffice\PhpPresentation\PhpPresentation;
 use PhpOffice\PhpPresentation\Shape\AutoShape;
-use PhpOffice\PhpPresentation\Style\Color;
-use PhpOffice\PhpPresentation\Style\Fill;
 use PhpOffice\PhpPresentation\Style\Outline;
-use PhpOffice\PhpPresentation\Writer\PowerPoint2007;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use ZipArchive;
 
 class AutoShapeTest extends TestCase
 {
@@ -75,51 +70,5 @@ class AutoShapeTest extends TestCase
     {
         $shape = new AutoShape();
         self::assertNull($shape->getRoundRectCorner());
-    }
-
-    public function testWriterEmitsAdjGuideForRoundRect(): void
-    {
-        $ppt = new PhpPresentation();
-        $slide = $ppt->getActiveSlide();
-
-        $width = 200;
-        $height = 100;
-        $padding = 5;
-        $minHalf = (int) floor(min($width, $height) / 2);
-        $expectedAdj = (int) round($padding / $minHalf * 50000); // 5000
-
-        $shape = (new AutoShape())
-            ->setType(AutoShape::TYPE_ROUNDED_RECTANGLE)
-            ->setWidth($width)->setHeight($height)
-            ->setRoundRectCorner($padding);
-
-        // Give it a fill so it's an obvious shape
-        $shape->getFill()->setFillType(Fill::FILL_SOLID)->setStartColor(new Color('FFFFFFFF'));
-        $slide->addShape($shape);
-
-        $tmpFile = tempnam(sys_get_temp_dir(), 'pptx_');
-        $writer = new PowerPoint2007($ppt);
-        $writer->save($tmpFile);
-
-        // Open the pptx and read slide1.xml
-        $zip = new ZipArchive();
-        self::assertTrue($zip->open($tmpFile) === true, 'Failed to open pptx zip');
-        $xml = $zip->getFromName('ppt/slides/slide1.xml');
-        $zip->close();
-        @unlink($tmpFile);
-
-        self::assertIsString($xml);
-
-        // Must contain roundRect geometry and the adj guide with expected value
-        self::assertStringContainsString('<a:prstGeom prst="roundRect">', $xml);
-
-        // fmla="val N" (there is a space after 'val' in writer)
-        self::assertSame(
-            1,
-            preg_match(
-                sprintf('/<a:gd[^>]+name="adj"[^>]+fmla="val %d"/', $expectedAdj),
-                (string) $xml
-            )
-        );
     }
 }
